+++ client/src/app/app.module.ts
@@ -9,9 +9,19 @@ import { SharedModule } from "./shared/shared.module";
 import { ChatModule } from "./chat/chat.module";
 import { TopbarComponent } from "./topbar/topbar.component";
 import { AccountSettingsComponent } from "./modals/account-settings/account-settings.component";
+import { LoginComponent } from "./login/login.component";
+import { HTTP_INTERCEPTORS, HttpClientModule } from "@angular/common/http";
+import { JwtInterceptor } from "./_helpers/jwt.interceptor";
+import { ErrorInterceptor } from "./_helpers/error.interceptor";
+import { fakeBackendProvider } from "./_helpers/fake_backend";
 
 @NgModule({
-  declarations: [AppComponent, TopbarComponent, AccountSettingsComponent],
+  declarations: [
+    AppComponent,
+    TopbarComponent,
+    AccountSettingsComponent,
+    LoginComponent
+  ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
@@ -19,9 +29,16 @@ import { AccountSettingsComponent } from "./modals/account-settings/account-sett
     SharedModule,
     ChatModule,
     FormsModule,
+    HttpClientModule,
     ReactiveFormsModule.withConfig({ warnOnNgModelWithFormControl: "never" })
   ],
-  providers: [],
+  providers: [
+    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+
+    // provider used to create fake backend
+    fakeBackendProvider
+  ],
   bootstrap: [AppComponent],
   entryComponents: [AccountSettingsComponent]
 })
